pyproject.toml
[project]
name = "stray-id"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "demiurg39", email = "chiryagov2014@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]

[project.scripts]
stray-id = "stray_id:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
